+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 # v1.0.0(2023-03-21)
 * Added target argument for command line. `m3u-filter -t <target_name> -t <target_name>`. Target names should be provided in the config.
+* Added filter to mapper definition.
+* Refactored filter parsing.
 
 # v0.9.9(2023-03-20)
 * Added optional 'enabled' property to input and target. Default is true.